Update dropboxlist.php
Introduced an item count, partially because it is useful to know, but mainly because there are issues with processing more than ~1000 items at a time.  When there are too many items, the collection and tag fields are not processed, with the result that all items are added with no collection and no tags.

I suspect this is due to the length limit of an HTTP POST request, and that the POST content is truncated.  The list of files to be processed comes before the other fields, and hence they drop off the bottom of the request.